[vpr] add get_parallel_segs.cpp to librrgraph
