+++ vpr/src/draw/draw.cpp
@@ -759,8 +759,6 @@ void act_on_mouse_press(ezgl::application* app, GdkEventButton* event, double x,
 void act_on_mouse_move(ezgl::application* app, GdkEventButton* event, double x, double y) {
     //  std::cout << "Mouse move at coordinates (" << x << "," << y << ") "<< std::endl;
 
-    const auto& device_ctx = g_vpr_ctx.device(); 
-
     // user has clicked the window button, in window mode
     if (window_point_1_collected) {
         // draw a grey, dashed-line box to indicate the zoom-in region
@@ -782,6 +780,7 @@ void act_on_mouse_move(ezgl::application* app, GdkEventButton* event, double x, 
         if (hit_node != OPEN) {
             //Update message
 
+            const auto& device_ctx = g_vpr_ctx.device();
             std::string info = describe_rr_node(device_ctx.rr_graph, device_ctx.grid, device_ctx.rr_indexed_data, hit_node);
             std::string msg = vtr::string_fmt("Moused over %s", info.c_str());
             app->update_message(msg.c_str());
