+++ vpr/src/draw/search_bar.cpp
@@ -136,12 +136,12 @@ void search_and_highlight(GtkWidget* /*widget*/, ezgl::application* app) {
 }
 
 bool highlight_rr_nodes(int hit_node) {
-    const auto& device_ctx = g_vpr_ctx.device(); 
     t_draw_state* draw_state = get_draw_state_vars();
 
     char message[250] = "";
 
     if (hit_node != OPEN) {
+        const auto& device_ctx = g_vpr_ctx.device(); 
         auto nodes = draw_expand_non_configurable_rr_nodes(hit_node);
         for (auto node : nodes) {
             if (draw_state->draw_rr_node[node].color != ezgl::MAGENTA) {
